--- .jsbeautifyrc
@@ -1,6 +1,0 @@
-{
-  "html": {
-    "indent_size": 2,
-    "unformatted": ["svg", "pre"]
-  }
-}
