--- README.md
@@ -1,2 +1,0 @@
-# hc_practice  
-課題提出用のリポジトリです
