--- common/scss/style.scss
@@ -1,84 +1,0 @@
-.l-main {
-  display: grid;
-  gap: 0 24px;
-  grid-template-columns:
-    1fr
-    minmax(375px - (24px * 2), 1000px)
-    1fr;
-
-  &__inner {
-    grid-column: 2/3;
-
-    >:first-child {
-      margin-top: 0!important;
-    }
-
-    >:last-child {
-      margin-bottom: 0!important;
-    }
-  }
-}
-
-.m-filter {
-  border: 1px solid #333;
-  display: flex;
-  flex-direction: row;
-  gap: 8px;
-  padding: 16px 24px;
-  justify-content: space-around;
-  flex-wrap: wrap;
-  margin-bottom: 32px;
-
-
-  li {
-    list-style: none;
-    display: flex;
-    column-gap: 4px;
-    flex-direction: row;
-    align-items: baseline;
-    position: relative;
-  }
-
-  input {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-    border: 1px solid #333;
-    border-radius: 50%;
-    width: 1em;
-    height: 1em;
-
-    &:checked::before {
-      
-    }
-  }
-}
-
-.m-panel {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 24px;
-  margin-bottom: 32px;
-
-  &__item {
-    border: 1px solid #f1f2f6;
-    display: flex;
-    flex-direction: column-reverse;
-  }
-
-  &__media {
-    overflow: hidden;
-  }
-
-  &__image {
-    aspect-ratio: 16/9;
-    height: 100%;
-    width: 100%;
-    object-fit: cover;
-    object-position: center;
-  }
-
-  &__contnet {
-    padding: 16px;
-  }
-}
