+++ index.html
@@ -5,8 +5,8 @@
 <link rel="icon" type="image/svg+xml" href="/vite.svg" />
 <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 <title></title>
-<link rel="stylesheet" href="common/scss/destyle.scss">
-<link rel="stylesheet" href="common/scss/style.scss">
+<link rel="stylesheet" href="common/css/destyle.css">
+<link rel="stylesheet" href="common/css/style.css">
 <script src="common/js/main.js" defer></script>
 </head>
 
@@ -21,7 +21,7 @@
 </ul><!-- /.m-filter -->
 
 
-<ul class="m-panel" id="js-panel">
+<ul class="m-panel" id="js-panel"></ul>
 
 
 <template class="js-panel__template">
@@ -43,9 +43,6 @@
 <template class="m-panel__templateError">
 <li>Not found</li>
 </template>
-
-
-</ul><!-- /.m-panel -->
 </div>
 </main><!-- /.l-main -->
 </body>
