--- postcss.config.cjs
@@ -1,13 +1,0 @@
-module.exports = {
-  plugins: {
-    'postcss-normalize-charset': {},
-    autoprefixer: {},
-    'postcss-sort-media-queries': {},
-    'css-declaration-sorter':{order:'smacss'},
-    '@fullhuman/postcss-purgecss': {
-      content: ['./src/**/*.html','./src/js/**/*.js'],
-      //除外設定　https://purgecss.com/safelisting.html
-      safelist: ['hoge']
-    },
-  }
-}
