--- vite.config.js
@@ -1,32 +1,0 @@
-import { defineConfig } from 'vite';
-
-export default defineConfig({
-  root: './src',
-  build: {
-    outDir: '../dist',
-    emptyOutDir: true,
-    cssCodeSplit: true,
-    rollupOptions: {
-      output: {
-        assetFileNames: (assetInfo) => {
-          let extType = assetInfo.name.split('.')[1];
-          if (/ttf|otf|eot|woff|woff2/i.test(extType)) {
-            extType = 'fonts';
-          }
-          if (/png|jpe?g|svg|gif|tiff|bmp|ico/i.test(extType)) {
-            extType = 'images';
-          }
-          if(extType === 'css') {
-            return `assets/css/style.css`;
-          }
-          return `assets/${extType}/[name][extname]`;
-        },
-        chunkFileNames: 'assets/js/[name].js',
-        entryFileNames: 'assets/js/[name].js',
-      },
-    },
-  },
-  server: {
-    host: true,
-  },
-});
